feat: prepare sell evolution
