+++ libs/ledger-live-common/src/exchange/platform/startExchange.ts
@@ -5,7 +5,7 @@ import { ExchangeTypes, createExchange, isExchangeTypeNg } from "@ledgerhq/hw-ap
 import { withDevice } from "../../hw/deviceAccess";
 import type { ExchangeRequestEvent } from "../../hw/actions/startExchange";
 import { StartExchangeInput } from "./types";
-import { getProviderConfig } from "../swap";
+import { getSwapProvider } from "../providers";
 
 const withDevicePromise = (deviceId, fn) =>
   firstValueFrom(withDevice(deviceId)(transport => from(fn(transport))));
@@ -20,7 +20,7 @@ const startExchange = (input: StartExchangeInput): Observable<ExchangeRequestEve
         let version;
         if (unsubscribed) return;
         if (provider) {
-          const providerConfig = getProviderConfig(provider);
+          const providerConfig = getSwapProvider(provider);
           if (providerConfig.type !== "CEX") {
             throw new Error(`Unsupported provider type ${providerConfig.type}`);
           }
