+++ libs/ledger-live-common/src/exchange/swap/getExchangeRates.ts
@@ -8,10 +8,11 @@ import {
   SwapExchangeRateAmountTooLow,
   SwapExchangeRateAmountTooLowOrTooHigh,
 } from "../../errors";
-import { getProviderConfig, getSwapAPIBaseURL, getSwapAPIError } from "./";
+import { getSwapAPIBaseURL, getSwapAPIError } from "./";
 import { mockGetExchangeRates } from "./mock";
 import type { CustomMinOrMaxError, GetExchangeRates } from "./types";
 import { isIntegrationTestEnv } from "./utils/isIntegrationTestEnv";
+import { getSwapProvider } from "../providers";
 
 const getExchangeRates: GetExchangeRates = async ({
   exchange,
@@ -137,7 +138,7 @@ const inferError = (
   const tenPowMagnitude = new BigNumber(10).pow(unitFrom.magnitude);
   const { amountTo, minAmountFrom, maxAmountFrom, errorCode, errorMessage, provider, status } =
     responseData;
-  const isDex = getProviderConfig(provider).type === "DEX";
+  const isDex = getSwapProvider(provider).type === "DEX";
 
   // DEX quotes are out of limits error. We do not know if it is a low or high limit, neither the amount.
   if ((!minAmountFrom || !maxAmountFrom) && status === "error" && errorCode !== 300 && isDex) {
